# backend-go/openapi.yaml
openapi: 3.0.3
info:
  title: Cancer Detection Backend API
  version: 1.0.0
  description: |
    Gateway-facing backend API.
    Internal microservice communication may use gRPC, but external clients should use these HTTP endpoints.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/detections:
    get:
      tags: [Detections]
      summary: List detections
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: priority
          schema: { type: string, enum: [HIGH, MEDIUM, LOW] }
        - in: query
          name: orderByPriority
          schema: { type: boolean, default: true }
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  detections:
                    type: array
                    items: { type: object, additionalProperties: true }
                  nextCursor: { type: string, nullable: true }
  /api/v1/detections/{id}:
    get:
      tags: [Detections]
      summary: Get detection by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/detections/{id}/review:
    post:
      tags: [Detections]
      summary: Review detection (DOCTOR/ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectionReviewRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema: { type: string }
    LimitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          nullable: true
...
    DetectionReviewRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [REVIEWED, CONFIRMED]
        reviewNotes:
          type: string
        findings:
          type: object
          additionalProperties: true
...